+++ src/aqua/cli/diagnostic_config.py
@@ -92,6 +92,11 @@ diagnostic_config = {
             'config_file': 'config.yaml',
             'source_path': 'diagnostics/ocean3d/cli',
             'target_path': 'diagnostics/ocean3d/cli'
+        },
+            {
+            'config_file': 'config_ocean_drift.yaml',
+            'source_path': 'config/diagnostics/ocean3d',
+            'target_path': 'diagnostics/ocean3d'
         }
     ],
     'radiation': [